kp-#6: refactor code and create Users layout which controls UsersList and UserPage
